requirements.txt
SQLAlchemy==1.4.39
requests==2.28.1
pandas==1.4.3
beautifulsoup4==4.11.1
aiogram==2.21
python-dotenv==0.20.0
asyncpg==0.26.0